Updated game manager and got basic round based movement working
You can only move up to your class' AP in each round. By clicking on a square you will move to it, assuming you have enough AP to do so. If you have a path to a tile when you end your turn then you will spend the rest of your unused AP pathing towards that tile.

+++ Assets/Scripts/TileType.cs
@@ -1,11 +1,10 @@
 ﻿using UnityEngine;
 using System.Collections;
-
 [System.Serializable]
 public class TileType {
 	public string name;
 	public GameObject tilePrefab;
 	[Range(0, 100)] public int frequency; // Maximum of 100%
 	public bool isWalkable = true;
-	public float movementCost = 1;
+	public int movementCost = 1;
 }
